Split files
Split into Algorithms and Misc (Tree, Node, MS)
